Moves helper function `h` to base/browser/dom
